+++ rvp64_exec/builtin-rvp64-kaddw.c
@@ -9,7 +9,7 @@
 #include "test_utils.h"
 
 static __attribute__ ((noinline))
-int64_t kaddw (int32_t ra, int32_t rb)
+int32_t kaddw (int32_t ra, int32_t rb)
 {
   return __rv_kaddw (ra, rb);
 }
@@ -17,7 +17,7 @@ int64_t kaddw (int32_t ra, int32_t rb)
 int
 main ()
 {
-  int64_t sa = kaddw (0x11223344, 0x55667788);
+  int32_t sa = kaddw (0x11223344, 0x55667788);
 
   if (sa != 0x6688aacc)
     abort ();
