+++ rvp64_scan/builtin-rvp64-kaddh.c
@@ -9,7 +9,7 @@
 #include <stdint.h>
 
 static __attribute__ ((noinline))
-int64_t hddak (int32_t ra, int32_t rb)
+int32_t hddak (int32_t ra, int32_t rb)
 {
   return __rv_kaddh (ra, rb);
 }
